requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
faker==8.12.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
joblib==1.0.1
python-multipart==0.0.5
alembic==1.7.1
psycopg2-binary==2.9.1 